@@ -2,52 +2,55 @@
 
     <PropertyGroup>
         <OutputType>Exe</OutputType>
-        <TargetFramework>net8.0</TargetFramework>
-        <LangVersion>12</LangVersion>
+        <TargetFramework>net9.0</TargetFramework>
+        <LangVersion>latest</LangVersion>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
-      <UserSecretsId>77a428ac-67e2-4da2-8bf8-6a777dbf2d74</UserSecretsId>
-  </PropertyGroup>
+        <UserSecretsId>77a428ac-67e2-4da2-8bf8-6a777dbf2d74</UserSecretsId>
+    </PropertyGroup>
 
     <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
-      <DebugSymbols>true</DebugSymbols>
+        <DebugSymbols>true</DebugSymbols>
     </PropertyGroup>
-    
+
     <PropertyGroup>
         <FarkleGenerateHtml>true</FarkleGenerateHtml>
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="AdventOfCodeSupport" Version="2.3.2" />
-      <PackageReference Include="Farkle" Version="6.5.1" />
-      <PackageReference Include="Farkle.Tools.MSBuild" Version="6.5.1" />
-      <PackageReference Include="MathNet.Numerics" Version="5.0.0" />
-      <PackageReference Include="NetTopologySuite" Version="2.5.0" />
-      <PackageReference Include="QuikGraph" Version="2.5.0" />
-      <PackageReference Include="QuikGraph.Graphviz" Version="2.5.0" />
-      <PackageReference Include="Rationals" Version="2.3.0" />
-      <PackageReference Include="VDS.Common" Version="2.0.0" />
+        <PackageReference Include="AdventOfCodeSupport" Version="2.4.3" />
+        <PackageReference Include="Farkle" Version="6.5.2" />
+        <PackageReference Include="Farkle.Tools.MSBuild" Version="6.5.2">
+          <PrivateAssets>all</PrivateAssets>
+          <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+        </PackageReference>
+        <PackageReference Include="MathNet.Numerics" Version="5.0.0"/>
+        <PackageReference Include="NetTopologySuite" Version="2.5.0"/>
+        <PackageReference Include="QuikGraph" Version="2.5.0"/>
+        <PackageReference Include="QuikGraph.Graphviz" Version="2.5.0"/>
+        <PackageReference Include="Rationals" Version="2.3.0"/>
+        <PackageReference Include="VDS.Common" Version="2.0.0"/>
     </ItemGroup>
 
     <ItemGroup>
-      <Compile Remove="2023\Inputs\**" />
+        <Compile Remove="2023\Inputs\**"/>
     </ItemGroup>
 
     <ItemGroup>
-      <EmbeddedResource Remove="2023\Inputs\**" />
+        <EmbeddedResource Remove="2023\Inputs\**"/>
     </ItemGroup>
 
     <ItemGroup>
-      <None Remove="2023\Inputs\**" />
-      <None Remove="Saved.json" />
+        <None Remove="2023\Inputs\**"/>
+        <None Remove="Saved.json"/>
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\AoC.Support\AoC.Support.csproj" />
+        <ProjectReference Include="..\AoC.Support\AoC.Support.csproj"/>
     </ItemGroup>
 
     <ItemGroup>
-      <Folder Include="2022\Inputs\" />
+        <Folder Include="2022\Inputs\"/>
     </ItemGroup>
 
 </Project>
