@@ -27,6 +27,47 @@ namespace AoC2023._2023;
 using Vertex = Vertex<int>;
 
 public class Day16 : Adventer {
+    private Problem problem;
+
+    public Day16() {
+        Bag["test"] =
+            """
+            .|...\....
+            |.-.\.....
+            .....|-...
+            ........|.
+            ..........
+            .........\
+            ..../.\\..
+            .-.-/..|..
+            .|....-|.\
+            ..//.|....
+            """;
+    }
+
+    protected override void InternalOnLoad() {
+        problem = new Problem(Input.Lines);
+    }
+
+    protected override object InternalPart1() {
+        // var tracer = new Problem.TraceObserver(problem);
+        var visited = new HashSet<Vertex>();
+        foreach (var photon in problem.Trace())
+            // tracer.Observe(photon);
+            visited.Add(photon.Position);
+        // var str = tracer.ToString();
+        // Console.WriteLine(str);
+        // Console.WriteLine();
+        return visited.Count;
+    }
+
+    protected override object InternalPart2() {
+        return problem.PossibleStarts()
+            .AsParallel()
+            .Select(p => problem.Energized(p).Count)
+            .Max();
+    }
+
     public readonly record struct Photon(Vertex Position, Direction Direction) {
         public Photon Move() {
             return this with { Position = Position + Direction.ToVertex<int>() };
@@ -43,50 +84,50 @@ public class Day16 : Adventer {
 
     public abstract record Tile {
         public abstract IEnumerable<Direction> Transit(Direction incoming);
-        
+
         public abstract override string ToString();
     }
 
     public sealed record Mirror : Tile {
         public enum Orientation {
             Forward = 0,
-            Backward = 1,
+            Backward = 1
         }
+
+        public static readonly Mirror Forward = new(Orientation.Forward);
+        public static readonly Mirror Backward = new(Orientation.Backward);
+
+        private static readonly Direction[,] TransitTable = CreateTransitTable();
 
         private Mirror(Orientation Kind) {
             this.Kind = Kind;
         }
 
-        public static readonly Mirror Forward = new Mirror(Orientation.Forward);
-        public static readonly Mirror Backward = new Mirror(Orientation.Backward);
-
         public Orientation Kind { get; }
 
         private static Direction[,] CreateTransitTable() {
             var table = new Direction[2, 4];
-            table[(int) Orientation.Forward, (int) Direction.Up] = Direction.Right;
-            table[(int) Orientation.Forward, (int) Direction.Down] = Direction.Left;
-            table[(int) Orientation.Forward, (int) Direction.Left] = Direction.Down;
-            table[(int) Orientation.Forward, (int) Direction.Right] = Direction.Up;
-            table[(int) Orientation.Backward, (int) Direction.Up] = Direction.Left;
-            table[(int) Orientation.Backward, (int) Direction.Down] = Direction.Right;
-            table[(int) Orientation.Backward, (int) Direction.Left] = Direction.Up;
-            table[(int) Orientation.Backward, (int) Direction.Right] = Direction.Down;
+            table[(int)Orientation.Forward, (int)Direction.Up] = Direction.Right;
+            table[(int)Orientation.Forward, (int)Direction.Down] = Direction.Left;
+            table[(int)Orientation.Forward, (int)Direction.Left] = Direction.Down;
+            table[(int)Orientation.Forward, (int)Direction.Right] = Direction.Up;
+            table[(int)Orientation.Backward, (int)Direction.Up] = Direction.Left;
+            table[(int)Orientation.Backward, (int)Direction.Down] = Direction.Right;
+            table[(int)Orientation.Backward, (int)Direction.Left] = Direction.Up;
+            table[(int)Orientation.Backward, (int)Direction.Right] = Direction.Down;
 
             return table;
         }
 
-        private static readonly Direction[,] TransitTable = CreateTransitTable();
-
         public override Direction[] Transit(Direction incoming) {
-            return new[] { TransitTable[(int) Kind, (int) incoming] };
+            return new[] { TransitTable[(int)Kind, (int)incoming] };
         }
-        
+
         public override string ToString() {
             return Kind switch {
                 Orientation.Forward => "/",
                 Orientation.Backward => "\\",
-                _ => throw new ArgumentOutOfRangeException(nameof(Kind), Kind, "Invalid orientation"),
+                _ => throw new ArgumentOutOfRangeException(nameof(Kind), Kind, "Invalid orientation")
             };
         }
     }
@@ -94,12 +135,19 @@ public class Day16 : Adventer {
     public sealed record Splitter : Tile {
         public enum Orientation {
             UpDown = 0,
-            LeftRight = 1,
+            LeftRight = 1
         }
+
+        private static readonly Direction[][][] TransitTable = CreateTransitTable();
+
+        public static readonly Splitter UpDown = new(Orientation.UpDown);
+        public static readonly Splitter LeftRight = new(Orientation.LeftRight);
 
         private Splitter(Orientation kind) {
             Kind = kind;
         }
+
+        public Orientation Kind { get; }
 
         private static Direction[][][] CreateTransitTable() {
             var lrSplit = new[] { Direction.Left, Direction.Right };
@@ -111,78 +159,56 @@ public class Day16 : Adventer {
             var table = new[] {
                 new[] {
                     // UpDown
-                    up,      // Up
-                    down,    // Down
+                    up, // Up
+                    down, // Down
                     udSplit, // Left
-                    udSplit, // Right
+                    udSplit // Right
                 },
                 new[] {
                     // LeftRight
                     lrSplit, // Up
                     lrSplit, // Down
-                    left,    // Left
-                    right,   // Right
-                },
+                    left, // Left
+                    right // Right
+                }
             };
             return table;
         }
 
-        private static readonly Direction[][][] TransitTable = CreateTransitTable();
-
-        public Orientation Kind { get; }
-
-        public static readonly Splitter UpDown = new Splitter(Orientation.UpDown);
-        public static readonly Splitter LeftRight = new Splitter(Orientation.LeftRight);
-
 
         public override IEnumerable<Direction> Transit(Direction incoming) {
-            return TransitTable[(int) Kind][(int) incoming];
+            return TransitTable[(int)Kind][(int)incoming];
         }
-        
+
         public override string ToString() {
             return Kind switch {
                 Orientation.UpDown => "|",
                 Orientation.LeftRight => "-",
-                _ => throw new ArgumentOutOfRangeException(nameof(Kind), Kind, "Invalid orientation"),
+                _ => throw new ArgumentOutOfRangeException(nameof(Kind), Kind, "Invalid orientation")
             };
         }
     }
 
     public sealed record Empty : Tile {
-        public static readonly Empty Instance = new Empty();
+        public static readonly Empty Instance = new();
 
         private static readonly Direction[][] TransitTable = {
             new[] { Direction.Up },
             new[] { Direction.Down },
             new[] { Direction.Left },
-            new[] { Direction.Right },
+            new[] { Direction.Right }
         };
 
         public override IEnumerable<Direction> Transit(Direction incoming) {
-            return TransitTable[(int) incoming];
+            return TransitTable[(int)incoming];
         }
-        
+
         public override string ToString() {
             return ".";
         }
     }
 
     public class Problem {
-        private Dictionary<Vertex, Tile> Tiles { get; } = new Dictionary<Vertex, Tile>();
-        public int Height { get; }
-        public int Width { get; }
-
-        public static Tile ParseTile(char c) {
-            return c switch {
-                '/'  => Mirror.Forward,
-                '\\' => Mirror.Backward,
-                '|'  => Splitter.UpDown,
-                '-'  => Splitter.LeftRight,
-                '.'  => Empty.Instance,
-                _    => throw new ArgumentOutOfRangeException(nameof(c), c, "Invalid tile"),
-            };
-        }
-
         public Problem(string[] input) {
             Height = input.Length;
             Width = input[0].Length;
@@ -190,13 +216,26 @@ public class Day16 : Adventer {
                 var row = input[y];
                 for (var x = 0; x < Width; x++) {
                     var c = row[x];
-                    if (c == '.') {
-                        continue;
-                    }
+                    if (c == '.') continue;
 
                     Tiles.Add(new Vertex(x, y), ParseTile(c));
                 }
             }
+        }
+
+        private Dictionary<Vertex, Tile> Tiles { get; } = new();
+        public int Height { get; }
+        public int Width { get; }
+
+        public static Tile ParseTile(char c) {
+            return c switch {
+                '/' => Mirror.Forward,
+                '\\' => Mirror.Backward,
+                '|' => Splitter.UpDown,
+                '-' => Splitter.LeftRight,
+                '.' => Empty.Instance,
+                _ => throw new ArgumentOutOfRangeException(nameof(c), c, "Invalid tile")
+            };
         }
 
         private Tile GetTile(Vertex vertex) {
@@ -209,18 +248,14 @@ public class Day16 : Adventer {
             frontier.Enqueue(photon);
             while (frontier.Count > 0) {
                 var current = frontier.Dequeue();
-                if (!visited.Add(current)) {
-                    continue;
-                }
+                if (!visited.Add(current)) continue;
 
                 yield return current;
 
                 var tile = GetTile(current.Position);
                 foreach (var direction in tile.Transit(current.Direction)) {
                     var next = current.Move(direction);
-                    if (next.Position.ExistsInGrid(Width, Height)) {
-                        frontier.Enqueue(next);
-                    }
+                    if (next.Position.ExistsInGrid(Width, Height)) frontier.Enqueue(next);
                 }
             }
         }
@@ -236,9 +271,7 @@ public class Day16 : Adventer {
 
         public ISet<Vertex> Energized(Photon start) {
             var set = new HashSet<Vertex>();
-            foreach (var photon in Trace(start)) {
-                set.Add(photon.Position);
-            }
+            foreach (var photon in Trace(start)) set.Add(photon.Position);
 
             return set;
         }
@@ -268,8 +301,7 @@ public class Day16 : Adventer {
         }
 
         public class TraceObserver {
-            private Problem Problem { get; }
-            private Dictionary<Vertex, HashSet<Direction>> Transits { get; } = new Dictionary<Vertex, HashSet<Direction>>();
+            private readonly ImmutableArray<ImmutableArray<char>> baseGrid;
 
             public TraceObserver(Problem problem) {
                 Problem = problem;
@@ -281,17 +313,18 @@ public class Day16 : Adventer {
                         var tile = problem.GetTile(vertex);
                         lineBuilder.Add(tile.ToString()[0]);
                     }
+
                     builder.Add(lineBuilder.ToImmutable());
                 }
+
                 baseGrid = builder.ToImmutable();
             }
 
-            private readonly ImmutableArray<ImmutableArray<char>> baseGrid;
+            private Problem Problem { get; }
+            private Dictionary<Vertex, HashSet<Direction>> Transits { get; } = new();
 
             private HashSet<Direction> GetTransits(Vertex location) {
-                if (Transits.TryGetValue(location, out var tr)) {
-                    return tr;
-                }
+                if (Transits.TryGetValue(location, out var tr)) return tr;
 
                 var transits = new HashSet<Direction>();
                 Transits.Add(location, transits);
@@ -317,10 +350,10 @@ public class Day16 : Adventer {
                         var transits = GetTransits(vertex);
                         var cnt = transits.Count;
                         line[x] = cnt switch {
-                            0            => '.',
-                            1            => transits.First().ToChar(),
-                            > 1 and < 10 => (char) (cnt + '0'),
-                            _            => '!',
+                            0 => '.',
+                            1 => transits.First().ToChar(),
+                            > 1 and < 10 => (char)(cnt + '0'),
+                            _ => '!'
                         };
                     }
 
@@ -330,48 +363,5 @@ public class Day16 : Adventer {
                 return sb.ToString().TrimEnd();
             }
         }
-    }
-
-    private Problem problem;
-
-    protected override void InternalOnLoad() {
-        problem = new Problem(Input.Lines);
-    }
-
-    protected override object InternalPart1() {
-        // var tracer = new Problem.TraceObserver(problem);
-        var visited = new HashSet<Vertex>();
-        foreach (var photon in problem.Trace()) {
-            // tracer.Observe(photon);
-            visited.Add(photon.Position);
-            // var str = tracer.ToString();
-            // Console.WriteLine(str);
-            // Console.WriteLine();
-        }
-
-        return visited.Count;
-    }
-
-    protected override object InternalPart2() {
-        return problem.PossibleStarts()
-            .AsParallel()
-            .Select(p => problem.Energized(p).Count)
-            .Max();
-    }
-
-    public Day16() {
-        Bag["test"] =
-            """
-            .|...\....
-            |.-.\.....
-            .....|-...
-            ........|.
-            ..........
-            .........\
-            ..../.\\..
-            .-.-/..|..
-            .|....-|.\
-            ..//.|....
-            """;
     }
 }
