@@ -1,4 +1,5 @@
 ﻿#region license
+
 // AoC2023 - AoC.Support.Test - VectorExtensionsTest.cs
 // Copyright (C) 2023 Nicholas
 // 
@@ -14,6 +15,7 @@
 // 
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 #endregion
 
 using AoC.Support.Numerics;
@@ -24,76 +26,63 @@ namespace AoC.Support.Test.Numerics;
 [TestFixture]
 [TestOf(typeof(VectorExtensions))]
 public class VectorExtensionsTest {
-
     [Test]
     public void PopCount(
         [Random(1, 256 + 1, 10)] int count
-        ) {
-
+    ) {
         var numbers = Enumerable.Range(0, count).Select(_ => TestContext.CurrentContext.Random.NextULong()).ToArray();
         var expected = numbers.Select(ulong.PopCount).Aggregate((a, b) => a + b);
         var actual = numbers.AsSpan().PopCount();
         Assert.That(actual, Is.EqualTo(expected));
-
     }
-    
+
     [Test]
     public void PopCountAvx2(
         [Random(1, 256 + 1, 10)] int count
-        ) {
-
+    ) {
         var numbers = Enumerable.Range(0, count).Select(_ => TestContext.CurrentContext.Random.NextULong()).ToArray();
         var expected = numbers.Select(ulong.PopCount).Aggregate((a, b) => a + b);
         var actual = ((ReadOnlySpan<byte>)numbers.AsSpan().AsBytes()).PopCountAvx2();
         Assert.That(actual, Is.EqualTo(expected));
-
     }
-    
+
     [Test]
     public void PopCountSse3s(
         [Random(1, 256 + 1, 10)] int count
     ) {
-
         var numbers = Enumerable.Range(0, count).Select(_ => TestContext.CurrentContext.Random.NextULong()).ToArray();
         var expected = numbers.Select(ulong.PopCount).Aggregate((a, b) => a + b);
         var actual = ((ReadOnlySpan<byte>)numbers.AsSpan().AsBytes()).PopCountSse3s();
         Assert.That(actual, Is.EqualTo(expected));
-
     }
-    
+
     [Test]
     public void PopCountFallback(
         [Random(1, 256 + 1, 10)] int count
     ) {
-
         var numbers = Enumerable.Range(0, count).Select(_ => TestContext.CurrentContext.Random.NextULong()).ToArray();
         var expected = numbers.Select(ulong.PopCount).Aggregate((a, b) => a + b);
         var actual = ((ReadOnlySpan<byte>)numbers.AsSpan().AsBytes()).BoringPopCount();
         Assert.That(actual, Is.EqualTo(expected));
-
     }
-    
+
     [Test]
     public void PopCountAvx2Alternate(
         [Random(1, 256 + 1, 10)] int count
     ) {
-
         var numbers = Enumerable.Range(0, count).Select(_ => TestContext.CurrentContext.Random.NextULong()).ToArray();
         var expected = numbers.Select(ulong.PopCount).Aggregate((a, b) => a + b);
         var actual = ((ReadOnlySpan<byte>)numbers.AsSpan().AsBytes()).PopCountAvx2Alternate();
         Assert.That(actual, Is.EqualTo(expected));
-
     }
-    
+
     [Test]
     public void PopCountAvx2Unroll(
         [Random(1, 256 + 1, 10)] int count
     ) {
-
         var numbers = Enumerable.Range(0, count).Select(_ => TestContext.CurrentContext.Random.NextULong()).ToArray();
         var expected = numbers.Select(ulong.PopCount).Aggregate((a, b) => a + b);
         var actual = ((ReadOnlySpan<byte>)numbers.AsSpan().AsBytes()).PopCountAvx2ManualUnroll();
         Assert.That(actual, Is.EqualTo(expected));
-
     }
 }
