@@ -18,14 +18,11 @@
 
 #endregion
 
-using System.Collections;
 using System.Collections.Immutable;
 using System.Diagnostics.CodeAnalysis;
 using AoC.Support.Collections;
 using AoC.Support.Graphs;
 using QuikGraph;
-using QuikGraph.Algorithms;
-using QuikGraph.Algorithms.Services;
 using QuikGraph.Graphviz;
 using QuikGraph.Graphviz.Dot;
 
@@ -38,100 +35,11 @@ public class LongestSimplePathAlgorithm<TVertex, TEdge, TGraph>
     private readonly IEqualityComparer<TVertex> comparer;
     private readonly Func<IEdge<TVertex>, double> cost;
     private readonly Func<TVertex, TVertex, TEdge> edgeFactory;
+    private readonly KnownPopulationSet<TVertex>.Generator vertexSetGenerator;
+    private BidirectionalMatrixPartialGraph<TVertex, TEdge>? dominatorTree;
+    private IDictionary<TVertex, KnownPopulationSet<TVertex>> reachableFrom;
 
     private BidirectionalMatrixPartialGraph<TVertex, TEdge>? transitiveClosure;
-    private IDictionary<TVertex, KnownPopulationSet<TVertex>> reachableFrom;
-    private BidirectionalMatrixPartialGraph<TVertex, TEdge>? dominatorTree;
-    private readonly KnownPopulationSet<TVertex>.Generator vertexSetGenerator;
-
-    private class PathSegment : IEquatable<PathSegment> {
-        public PathSegment(TVertex vertex,
-            KnownPopulationSet<TVertex> visited,
-            ImmutableQueue<TVertex> path,
-            KnownPopulationSet<TVertex> reachableFromVertex,
-            IEqualityComparer<TVertex> comparer) {
-            Vertex = vertex;
-            Visited = visited;
-            Path = path;
-            var localKnown = reachableFromVertex.Clone();
-            localKnown.ExceptWith(Visited);
-            localKnown.Remove(vertex);
-            Reachable = localKnown;
-            Comparer = comparer;
-        }
-
-        public TVertex Vertex { get; }
-        public KnownPopulationSet<TVertex> Visited { get; }
-        public ImmutableQueue<TVertex> Path { get; }
-        public KnownPopulationSet<TVertex> Reachable { get; }
-
-        public IEqualityComparer<TVertex> Comparer { get; }
-
-        public bool Visits(TVertex v) => Visited.Contains(v) || Comparer.Equals(Vertex, v);
-        public int Length => Visited.Count + 1;
-
-        public bool Symmetric(PathSegment other) {
-            return Comparer.Equals(other.Vertex, Vertex) && Visited.SetEquals(other.Visited);
-        }
-
-        public override bool Equals(object? obj) {
-            return obj is PathSegment other && Equals(other);
-        }
-
-        public bool Equals(PathSegment? other) {
-            if (ReferenceEquals(null, other)) {
-                return false;
-            }
-
-            if (ReferenceEquals(this, other)) {
-                return true;
-            }
-
-            return Symmetric(other);
-        }
-
-        public bool Dominates(PathSegment other, bool proper = false) {
-            if (!Comparer.Equals(Vertex, other.Vertex)) {
-                return false;
-            }
-
-            if (Length < other.Length) {
-                return false;
-            }
-
-            if (proper) {
-                return Reachable.IsProperSupersetOf(other.Reachable);
-            }
-
-            return Reachable.IsSupersetOf(other.Reachable);
-        }
-
-        public override int GetHashCode() {
-            return HashCode.Combine(Vertex, Visited);
-        }
-
-        public static bool operator ==(PathSegment? left, PathSegment? right) {
-            return Equals(left, right);
-        }
-
-        public static bool operator !=(PathSegment? left, PathSegment? right) {
-            return !Equals(left, right);
-        }
-
-        public IEnumerable<TVertex> EnumeratePath() {
-            return Path.Append(Vertex);
-        }
-    }
-
-    private class PathState {
-        public TVertex Current { get; }
-        public ImmutableHashSet<TVertex> Reachable { get; }
-
-        public PathState(TVertex current, ImmutableHashSet<TVertex> reachable) {
-            Current = current;
-            Reachable = reachable;
-        }
-    }
 
     public LongestSimplePathAlgorithm(TGraph visitedGraph,
         TVertex root,
@@ -157,9 +65,7 @@ public class LongestSimplePathAlgorithm<TVertex, TEdge, TGraph>
     private BidirectionalMatrixPartialGraph<TVertex, TEdge> DominatorTree {
         [MemberNotNull(nameof(dominatorTree))]
         get {
-            if (dominatorTree != null) {
-                return dominatorTree;
-            }
+            if (dominatorTree != null) return dominatorTree;
 
             // var alg = NaiveDominatorSearchAlgorithm.Create(Graph,  Root, edgeFactory, comparer);
             var alg = DataFlowDominatorSearchAlgorithm.Create(Graph, Root, edgeFactory, comparer);
@@ -167,18 +73,22 @@ public class LongestSimplePathAlgorithm<TVertex, TEdge, TGraph>
             dominatorTree = alg.DominatorTree;
             var renderer = new GraphvizAlgorithm<TVertex, TEdge>(dominatorTree);
             renderer.FormatVertex += (sender, args) => {
-                if (Root.Equals(args.Vertex)) {
-                    args.VertexFormat.FillColor = GraphvizColor.Lavender;
-                }
+                if (Root.Equals(args.Vertex)) args.VertexFormat.FillColor = GraphvizColor.Lavender;
 
-                if (Dest.Equals(args.Vertex)) {
-                    args.VertexFormat.FillColor = GraphvizColor.LightPink;
-                }
+                if (Dest.Equals(args.Vertex)) args.VertexFormat.FillColor = GraphvizColor.LightPink;
 
                 args.VertexFormat.Label = args.Vertex.ToString();
             };
             File.WriteAllText($"{Graph.VertexCount}-{DateTime.Now.ToFileTimeUtc()}.dot", renderer.Generate());
             return dominatorTree;
+        }
+    }
+
+    private BidirectionalMatrixPartialGraph<TVertex, TEdge> TransitiveClosure {
+        get {
+            if (transitiveClosure == null) CalculateReachableFrom();
+
+            return transitiveClosure;
         }
     }
 
@@ -193,9 +103,8 @@ public class LongestSimplePathAlgorithm<TVertex, TEdge, TGraph>
 
     private List<TVertex> LongestPathBetween(TVertex source, TVertex dest) {
         Console.WriteLine($"Calculating longest path between {source} and {dest}");
-        if (!TransitiveClosure.ContainsEdge(source, dest)) {
+        if (!TransitiveClosure.ContainsEdge(source, dest))
             throw new ArgumentException("No path between source and dest");
-        }
 
         // var dag = DominatorTree.ShortestPathsDag(_ => 1, source);
         // if (!dag.Invoke(dest, out var path)) {
@@ -209,7 +118,8 @@ public class LongestSimplePathAlgorithm<TVertex, TEdge, TGraph>
         // var pairs = source.Once().Concat(dominatorPath).Zip(dominatorPath);
         var pairs = new[] { (Root, Dest) };
         foreach (var (cur, next) in pairs) {
-            var nextStep = LongestPathBetweenDominators(longestPath.Take(longestPath.Count - 1), visitedNodes, cur, next);
+            var nextStep =
+                LongestPathBetweenDominators(longestPath.Take(longestPath.Count - 1), visitedNodes, cur, next);
             longestPath.AddRange(nextStep.Skip(longestPath.Count));
             visitedNodes.UnionWith(longestPath);
         }
@@ -234,7 +144,8 @@ public class LongestSimplePathAlgorithm<TVertex, TEdge, TGraph>
         // var frontier = new Stack<PathSegment>();
         var dominant = new Dictionary<TVertex, PathSegment>(comparer);
         var curPathQueue = ImmutableQueue.CreateRange(currentPath);
-        var origin = new PathSegment(source, vertexSetGenerator.CreateSet(curPathQueue), curPathQueue, reachableFrom[source], comparer);
+        var origin = new PathSegment(source, vertexSetGenerator.CreateSet(curPathQueue), curPathQueue,
+            reachableFrom[source], comparer);
         frontier.Enqueue(origin, origin.Reachable.Count);
         // frontier.Push(origin);
         var cnt = 0;
@@ -243,10 +154,12 @@ public class LongestSimplePathAlgorithm<TVertex, TEdge, TGraph>
             cnt++;
             if (cnt % 5000 == 0) {
                 var curCnt = frontier.Count;
-                Console.WriteLine($"Frontier size at {cnt}: {frontier.Count}, probed: {probed}, rejected: {probed - cnt}");
+                Console.WriteLine(
+                    $"Frontier size at {cnt}: {frontier.Count}, probed: {probed}, rejected: {probed - cnt}");
                 if (dominant.TryGetValue(dest, out var ps)) {
                     Console.WriteLine($"Best path so far of length {ps.Length}");
-                } else {
+                }
+                else {
                     Console.WriteLine("No path candidate found yet");
                     var best = frontier.Peek();
                     Console.WriteLine($"Best path being investigated ends at {best.Vertex} of length {best.Length}");
@@ -272,35 +185,23 @@ public class LongestSimplePathAlgorithm<TVertex, TEdge, TGraph>
             // var state = frontier.Pop();
 
             if (!dominant.TryAdd(state.Vertex, state)) {
-                if (dominant[state.Vertex].Dominates(state)) {
-                    continue;
-                }
+                if (dominant[state.Vertex].Dominates(state)) continue;
 
                 dominant[state.Vertex] = state;
             }
 
             // We've already visited the destination, no point in continuing.
-            if (state.Visits(dest)) {
-                continue;
-            }
+            if (state.Visits(dest)) continue;
 
             foreach (var e in Graph.OutEdges(state.Vertex)) {
                 probed++;
-                if (prevPath.Contains(e.Target)) {
-                    continue;
-                }
+                if (prevPath.Contains(e.Target)) continue;
 
-                if (state.Visits(e.Target)) {
-                    continue;
-                }
+                if (state.Visits(e.Target)) continue;
 
-                if (!CanReachDestFrom(e.Target)) {
-                    continue;
-                }
+                if (!CanReachDestFrom(e.Target)) continue;
 
-                if (!comparer.Equals(e.Target, dest) && !TransitiveClosure.ContainsEdge(e.Target, dest)) {
-                    continue;
-                }
+                if (!comparer.Equals(e.Target, dest) && !TransitiveClosure.ContainsEdge(e.Target, dest)) continue;
 
                 var newVisited = state.Visited.Clone();
                 newVisited.Add(state.Vertex);
@@ -312,29 +213,23 @@ public class LongestSimplePathAlgorithm<TVertex, TEdge, TGraph>
                     comparer
                 );
 
-                if (dominant.TryGetValue(newState.Vertex, out var curDom)) {
-                    if (curDom.Dominates(newState)) {
+                if (dominant.TryGetValue(newState.Vertex, out var curDom))
+                    if (curDom.Dominates(newState))
                         continue;
-                    }
-                }
 
                 frontier.Enqueue(newState, newState.Reachable.Count);
                 // frontier.Push(newState);
             }
         }
 
-        if (!dominant.TryGetValue(dest, out var res)) {
-            throw new InvalidOperationException("No path found");
-        }
+        if (!dominant.TryGetValue(dest, out var res)) throw new InvalidOperationException("No path found");
 
         return res.EnumeratePath();
     }
 
     [MemberNotNull(nameof(transitiveClosure), nameof(reachableFrom))]
     private void CalculateReachableFrom() {
-        if (transitiveClosure != null && reachableFrom != null) {
-            return;
-        }
+        if (transitiveClosure != null && reachableFrom != null) return;
 
         var subGraph = BidirectionalMatrixPartialGraph<TVertex, TEdge>.Create(Graph, edgeFactory, comparer);
         var tc = subGraph.ComputeTransitiveClosure();
@@ -343,9 +238,7 @@ public class LongestSimplePathAlgorithm<TVertex, TEdge, TGraph>
         reachableFrom = new Dictionary<TVertex, KnownPopulationSet<TVertex>>(comparer);
         foreach (var v in Graph.Vertices) {
             var reach = vertexSetGenerator.CreateSet();
-            if (subGraph.ContainsVertex(v)) {
-                reach.UnionWith(TransitiveClosure.OutEdges(v).Select(e => e.Target));
-            }
+            if (subGraph.ContainsVertex(v)) reach.UnionWith(TransitiveClosure.OutEdges(v).Select(e => e.Target));
 
             reachableFrom[v] = reach;
         }
@@ -360,14 +253,86 @@ public class LongestSimplePathAlgorithm<TVertex, TEdge, TGraph>
         return TransitiveClosure.ContainsEdge(v, Dest);
     }
 
-    private BidirectionalMatrixPartialGraph<TVertex, TEdge> TransitiveClosure {
-        get {
-            if (transitiveClosure == null) {
-                CalculateReachableFrom();
-            }
-
-            return transitiveClosure;
+    private class PathSegment : IEquatable<PathSegment> {
+        public PathSegment(TVertex vertex,
+            KnownPopulationSet<TVertex> visited,
+            ImmutableQueue<TVertex> path,
+            KnownPopulationSet<TVertex> reachableFromVertex,
+            IEqualityComparer<TVertex> comparer) {
+            Vertex = vertex;
+            Visited = visited;
+            Path = path;
+            var localKnown = reachableFromVertex.Clone();
+            localKnown.ExceptWith(Visited);
+            localKnown.Remove(vertex);
+            Reachable = localKnown;
+            Comparer = comparer;
         }
+
+        public TVertex Vertex { get; }
+        public KnownPopulationSet<TVertex> Visited { get; }
+        public ImmutableQueue<TVertex> Path { get; }
+        public KnownPopulationSet<TVertex> Reachable { get; }
+
+        public IEqualityComparer<TVertex> Comparer { get; }
+
+        public int Length => Visited.Count + 1;
+
+        public bool Equals(PathSegment? other) {
+            if (ReferenceEquals(null, other)) return false;
+
+            if (ReferenceEquals(this, other)) return true;
+
+            return Symmetric(other);
+        }
+
+        public bool Visits(TVertex v) {
+            return Visited.Contains(v) || Comparer.Equals(Vertex, v);
+        }
+
+        public bool Symmetric(PathSegment other) {
+            return Comparer.Equals(other.Vertex, Vertex) && Visited.SetEquals(other.Visited);
+        }
+
+        public override bool Equals(object? obj) {
+            return obj is PathSegment other && Equals(other);
+        }
+
+        public bool Dominates(PathSegment other, bool proper = false) {
+            if (!Comparer.Equals(Vertex, other.Vertex)) return false;
+
+            if (Length < other.Length) return false;
+
+            if (proper) return Reachable.IsProperSupersetOf(other.Reachable);
+
+            return Reachable.IsSupersetOf(other.Reachable);
+        }
+
+        public override int GetHashCode() {
+            return HashCode.Combine(Vertex, Visited);
+        }
+
+        public static bool operator ==(PathSegment? left, PathSegment? right) {
+            return Equals(left, right);
+        }
+
+        public static bool operator !=(PathSegment? left, PathSegment? right) {
+            return !Equals(left, right);
+        }
+
+        public IEnumerable<TVertex> EnumeratePath() {
+            return Path.Append(Vertex);
+        }
+    }
+
+    private class PathState {
+        public PathState(TVertex current, ImmutableHashSet<TVertex> reachable) {
+            Current = current;
+            Reachable = reachable;
+        }
+
+        public TVertex Current { get; }
+        public ImmutableHashSet<TVertex> Reachable { get; }
     }
 }
 
@@ -383,7 +348,8 @@ public static class LongestSimplePathAlgorithmExtensions {
         where TEdge : IEdge<TVertex>
         where TGraph : IBidirectionalGraph<TVertex, TEdge>
         where TVertex : notnull {
-        var alg = new LongestSimplePathAlgorithm<TVertex, TEdge, TGraph>(graph, root, dest, edgeFactory, comparer, cost);
+        var alg = new LongestSimplePathAlgorithm<TVertex, TEdge, TGraph>(graph, root, dest, edgeFactory, comparer,
+            cost);
         return alg.Compute();
     }
 }
