@@ -18,7 +18,6 @@
 
 #endregion
 
-using System.Diagnostics;
 using System.Numerics;
 using AoC.Support.Collections;
 using CommunityToolkit.HighPerformance;
@@ -28,16 +27,25 @@ namespace AoC.Support.Matrix;
 
 public class BitMatrix : ICloneable {
     private readonly BitArray[] data;
+
+    public BitMatrix(int rows, int columns, bool defaultValue = false) {
+        data = new BitArray[rows];
+        for (var i = 0; i < rows; i++) data[i] = new BitArray(columns, defaultValue);
+    }
+
+    public BitMatrix(Index2D size, bool defaultValue = false) : this(size.Y, size.X, defaultValue) { }
     public int RowCount => data.Length;
     public int ColumnCount => data[0].Count;
 
     public Index2D Size => new(RowCount, ColumnCount);
 
-    public BitMatrix(int rows, int columns, bool defaultValue = false) {
-        data = new BitArray[rows];
-        for (var i = 0; i < rows; i++) {
-            data[i] = new BitArray(columns, defaultValue);
-        }
+    public bool this[int row, int column] {
+        get => data[row][column];
+        set => data[row][column] = value;
+    }
+
+    object ICloneable.Clone() {
+        return Clone();
     }
 
     public static BitMatrix FromArray<TNumeric>(TNumeric[,] arr) where TNumeric : INumber<TNumeric> {
@@ -45,19 +53,10 @@ public class BitMatrix : ICloneable {
         var columns = arr.GetLength(1);
         var result = new BitMatrix(rows, columns);
         for (var i = 0; i < rows; i++) {
-            for (var j = 0; j < columns; j++) {
-                result[i, j] = arr[i, j] != TNumeric.Zero;
-            }
+            for (var j = 0; j < columns; j++) result[i, j] = arr[i, j] != TNumeric.Zero;
         }
-        
+
         return result;
-    }
-
-    public BitMatrix(Index2D size, bool defaultValue = false) : this(size.Y, size.X, defaultValue) { }
-
-    public bool this[int row, int column] {
-        get => data[row][column];
-        set => data[row][column] = value;
     }
 
     public BitArray GetRow(int row) {
@@ -66,25 +65,19 @@ public class BitMatrix : ICloneable {
 
     public BitArray GetColumn(int column) {
         var result = new BitArray(RowCount);
-        for (var i = 0; i < RowCount; i++) {
-            result[i] = data[i][column];
-        }
+        for (var i = 0; i < RowCount; i++) result[i] = data[i][column];
 
         return result;
     }
 
     public void CopyColumnInto(int column, BitArray bits) {
         ArgumentOutOfRangeException.ThrowIfNotEqual(bits.Count, RowCount);
-        for (var i = 0; i < RowCount; i++) {
-            bits[i] = data[i][column];
-        }
+        for (var i = 0; i < RowCount; i++) bits[i] = data[i][column];
     }
-    
+
     public void CopyRowInto(int row, BitArray bits) {
         ArgumentOutOfRangeException.ThrowIfNotEqual(bits.Count, ColumnCount);
-        for (var i = 0; i < ColumnCount; i++) {
-            bits[i] = data[row][i];
-        }
+        for (var i = 0; i < ColumnCount; i++) bits[i] = data[row][i];
     }
 
     public void SetRow(int row, BitArray bits) {
@@ -93,9 +86,7 @@ public class BitMatrix : ICloneable {
 
     public void SetColumn(int column, BitArray bits) {
         ArgumentOutOfRangeException.ThrowIfNotEqual(bits.Count, RowCount);
-        for (var i = 0; i < RowCount; i++) {
-            data[i][column] = bits[i];
-        }
+        for (var i = 0; i < RowCount; i++) data[i][column] = bits[i];
     }
 
     private bool RowValid(int row) {
@@ -107,73 +98,51 @@ public class BitMatrix : ICloneable {
     }
 
     public void SetRow(int row, bool value) {
-        if (!RowValid(row)) {
-            throw new ArgumentOutOfRangeException(nameof(row));
-        }
+        if (!RowValid(row)) throw new ArgumentOutOfRangeException(nameof(row));
 
         data[row].SetAll(value);
     }
 
     public void SetColumn(int column, bool value) {
-        if (!ColumnValid(column)) {
-            throw new ArgumentOutOfRangeException(nameof(column));
-        }
+        if (!ColumnValid(column)) throw new ArgumentOutOfRangeException(nameof(column));
 
-        for (var i = 0; i < RowCount; i++) {
-            data[i][column] = value;
-        }
+        for (var i = 0; i < RowCount; i++) data[i][column] = value;
     }
 
     public void SetAll(bool value) {
-        foreach (var row in data) {
-            row.SetAll(value);
-        }
+        foreach (var row in data) row.SetAll(value);
     }
 
     public BitMatrix And(BitMatrix other) {
         ArgumentOutOfRangeException.ThrowIfNotEqual(other.Size, Size);
-        for (var i = 0; i < RowCount; i++) {
-            data[i].And(other.data[i]);
-        }
+        for (var i = 0; i < RowCount; i++) data[i].And(other.data[i]);
 
         return this;
     }
 
     public BitMatrix Or(BitMatrix other) {
         ArgumentOutOfRangeException.ThrowIfNotEqual(other.Size, Size);
-        for (var i = 0; i < RowCount; i++) {
-            data[i].Or(other.data[i]);
-        }
+        for (var i = 0; i < RowCount; i++) data[i].Or(other.data[i]);
 
         return this;
     }
 
     public BitMatrix Xor(BitMatrix other) {
         ArgumentOutOfRangeException.ThrowIfNotEqual(other.Size, Size);
-        for (var i = 0; i < RowCount; i++) {
-            data[i].Xor(other.data[i]);
-        }
+        for (var i = 0; i < RowCount; i++) data[i].Xor(other.data[i]);
 
         return this;
     }
 
     public BitMatrix Not() {
-        for (var i = 0; i < RowCount; i++) {
-            data[i].Not();
-        }
+        for (var i = 0; i < RowCount; i++) data[i].Not();
 
         return this;
     }
 
-    object ICloneable.Clone() {
-        return Clone();
-    }
-
     public BitMatrix Clone() {
         var result = new BitMatrix(RowCount, ColumnCount);
-        for (var i = 0; i < RowCount; i++) {
-            result.data[i].Or(data[i]);
-        }
+        for (var i = 0; i < RowCount; i++) result.data[i].Or(data[i]);
 
         return result;
     }
@@ -181,9 +150,7 @@ public class BitMatrix : ICloneable {
     public BitMatrix Transposed() {
         var result = new BitMatrix(ColumnCount, RowCount);
         for (var i = 0; i < RowCount; i++) {
-            for (var j = 0; j < ColumnCount; j++) {
-                result[j, i] = this[i, j];
-            }
+            for (var j = 0; j < ColumnCount; j++) result[j, i] = this[i, j];
         }
 
         return result;
@@ -210,17 +177,13 @@ public class BitMatrix : ICloneable {
 
     public int CountSetBits() {
         var result = 0;
-        foreach (var row in data) {
-            result += row.CountSetBits();
-        }
+        foreach (var row in data) result += row.CountSetBits();
 
         return result;
     }
 
     public void Clear() {
-        foreach (var row in data) {
-            row.Clear();
-        }
+        foreach (var row in data) row.Clear();
     }
 
     public void ClearRow(int row) {
@@ -228,17 +191,13 @@ public class BitMatrix : ICloneable {
     }
 
     public void ClearColumn(int col) {
-        for (var i = 0; i < RowCount; i++) {
-            data[i][col] = false;
-        }
+        for (var i = 0; i < RowCount; i++) data[i][col] = false;
     }
-    
+
     public byte[,] ToBytes() {
         var result = new byte[RowCount, ColumnCount];
         for (var i = 0; i < RowCount; i++) {
-            for (var j = 0; j < ColumnCount; j++) {
-                result[i, j] = this[i, j].ToByte();
-            }
+            for (var j = 0; j < ColumnCount; j++) result[i, j] = this[i, j].ToByte();
         }
 
         return result;
@@ -246,9 +205,7 @@ public class BitMatrix : ICloneable {
 
     public IEnumerable<(int row, int column, bool value)> EnumerateIndexed() {
         for (var i = 0; i < RowCount; i++) {
-            for (var j = 0; j < ColumnCount; j++) {
-                yield return (i, j, this[i, j]);
-            }
+            for (var j = 0; j < ColumnCount; j++) yield return (i, j, this[i, j]);
         }
     }
 }
