@@ -1,4 +1,5 @@
 ﻿#region license
+
 // AoC2023 - AoC.Support - DenseRowMajorMatrixStorage.cs
 // Copyright (C) 2023 Nicholas
 // 
@@ -14,43 +15,45 @@
 // 
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 #endregion
 
 using CommunityToolkit.HighPerformance;
-using ILGPU.Runtime;
 using MathNet.Numerics.LinearAlgebra;
 using MathNet.Numerics.LinearAlgebra.Storage;
 
 namespace AoC.Support;
 
 public class DenseRowMajorMatrixStorage<T> : MatrixStorage<T> where T : struct, IEquatable<T>, IFormattable {
-    
     private readonly T[,] data;
-    
+
     public DenseRowMajorMatrixStorage(int rowCount, int columnCount, T v = default) : base(rowCount, columnCount) {
         data = new T[rowCount, columnCount];
-        if (v.Equals(default(T))) {
-            return;
-        }
+        if (v.Equals(default)) return;
 
         Span2D<T> span = data;
         span.Fill(v);
     }
-    
+
     private DenseRowMajorMatrixStorage(T[,] data) : base(data.GetLength(0), data.GetLength(1)) {
         this.data = data;
     }
-    
+
+    public override bool IsDense => true;
+    public override bool IsFullyMutable => true;
+
     public static DenseRowMajorMatrixStorage<T> OfStorage(MatrixStorage<T> matrix) {
         // using Clone skips initialization and just copies the data directly
         return new DenseRowMajorMatrixStorage<T>(matrix.AsArray().Clone() as T[,] ?? matrix.ToArray());
     }
-    
+
     public static DenseRowMajorMatrixStorage<T> OfMatrix(Matrix<T> matrix) {
         return OfStorage(matrix.Storage);
     }
 
-    public override bool IsMutableAt(int row, int column) => true;
+    public override bool IsMutableAt(int row, int column) {
+        return true;
+    }
 
     public override T At(int row, int column) {
         return data[row, column];
@@ -59,9 +62,6 @@ public class DenseRowMajorMatrixStorage<T> : MatrixStorage<T> where T : struct, 
     public override void At(int row, int column, T value) {
         data[row, column] = value;
     }
-
-    public override bool IsDense => true;
-    public override bool IsFullyMutable => true;
 
     public override void Clear() {
         Array.Clear(data);
@@ -77,13 +77,13 @@ public class DenseRowMajorMatrixStorage<T> : MatrixStorage<T> where T : struct, 
 }
 
 public static class DenseRowMajorMatrixStorage {
-    public static DenseRowMajorMatrixStorage<T> OfMatrix<T>(Matrix<T> matrix) where T : struct, IEquatable<T>, IFormattable {
+    public static DenseRowMajorMatrixStorage<T> OfMatrix<T>(Matrix<T> matrix)
+        where T : struct, IEquatable<T>, IFormattable {
         return DenseRowMajorMatrixStorage<T>.OfMatrix(matrix);
     }
-    
-    public static DenseRowMajorMatrixStorage<T> OfStorage<T>(MatrixStorage<T> matrix) where T : struct, IEquatable<T>, IFormattable {
+
+    public static DenseRowMajorMatrixStorage<T> OfStorage<T>(MatrixStorage<T> matrix)
+        where T : struct, IEquatable<T>, IFormattable {
         return DenseRowMajorMatrixStorage<T>.OfStorage(matrix);
     }
-    
-
 }
