@@ -1,4 +1,5 @@
 ﻿#region license
+
 // AoC2023 - AoC2023 - Day03.cs
 // Copyright (C) 2023 Nicholas
 // 
@@ -14,9 +15,9 @@
 // 
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
 #endregion
 
-using AdventOfCodeSupport;
 using AoC.Support;
 
 namespace AoC2023._2023;
@@ -24,6 +25,13 @@ namespace AoC2023._2023;
 using Vertex = Vertex<int>;
 
 public class Day03 : Adventer {
+    protected override object InternalPart1() {
+        throw new NotImplementedException();
+    }
+
+    protected override object InternalPart2() {
+        throw new NotImplementedException();
+    }
 
     public struct Number {
         public int Value { get; set; }
@@ -33,8 +41,6 @@ public class Day03 : Adventer {
 
         public Number(int value, int x, int y) {
             Value = value;
-            
-            
         }
 
         public override string ToString() {
@@ -49,17 +55,8 @@ public class Day03 : Adventer {
             }
         }
 
-        public int DecimalLength => (int) Math.Log10(Value) + 1;
-    }
-    public class Problem {
-        
-    }
-    
-    protected override object InternalPart1() {
-        throw new NotImplementedException();
+        public int DecimalLength => (int)Math.Log10(Value) + 1;
     }
 
-    protected override object InternalPart2() {
-        throw new NotImplementedException();
-    }
+    public class Problem { }
 }
