@@ -1,5 +1,4 @@
-﻿using System.Collections.Frozen;
-using System.Collections.Immutable;
+﻿using System.Collections.Immutable;
 using System.Diagnostics;
 
 namespace AoC2023._2023;
@@ -17,7 +16,7 @@ public enum Card {
     Jack,
     Queen,
     King,
-    Ace,
+    Ace
 }
 
 public enum AltCard {
@@ -33,7 +32,7 @@ public enum AltCard {
     Ten,
     Queen,
     King,
-    Ace,
+    Ace
 }
 
 public enum HandKind {
@@ -43,7 +42,7 @@ public enum HandKind {
     ThreeOfAKind,
     FullHouse,
     FourOfAKind,
-    FiveOfAKind,
+    FiveOfAKind
 }
 
 public static class Day07Extensions {
@@ -62,7 +61,7 @@ public static class Day07Extensions {
             "Q" => Card.Queen,
             "K" => Card.King,
             "A" => Card.Ace,
-            _   => throw new ArgumentException("Invalid card", nameof(input)),
+            _ => throw new ArgumentException("Invalid card", nameof(input))
         };
     }
 
@@ -81,62 +80,78 @@ public static class Day07Extensions {
             'Q' => Card.Queen,
             'K' => Card.King,
             'A' => Card.Ace,
-            _   => throw new ArgumentException("Invalid card", nameof(input)),
+            _ => throw new ArgumentException("Invalid card", nameof(input))
         };
     }
 
     public static AltCard ToAltCard(this Card card) {
         return card switch {
-            Card.Two   => AltCard.Two,
+            Card.Two => AltCard.Two,
             Card.Three => AltCard.Three,
-            Card.Four  => AltCard.Four,
-            Card.Five  => AltCard.Five,
-            Card.Six   => AltCard.Six,
+            Card.Four => AltCard.Four,
+            Card.Five => AltCard.Five,
+            Card.Six => AltCard.Six,
             Card.Seven => AltCard.Seven,
             Card.Eight => AltCard.Eight,
-            Card.Nine  => AltCard.Nine,
-            Card.Ten   => AltCard.Ten,
-            Card.Jack  => AltCard.Jack,
+            Card.Nine => AltCard.Nine,
+            Card.Ten => AltCard.Ten,
+            Card.Jack => AltCard.Jack,
             Card.Queen => AltCard.Queen,
-            Card.King  => AltCard.King,
-            Card.Ace   => AltCard.Ace,
-            _          => throw new ArgumentException("Invalid card", nameof(card)),
+            Card.King => AltCard.King,
+            Card.Ace => AltCard.Ace,
+            _ => throw new ArgumentException("Invalid card", nameof(card))
         };
     }
 }
 
 public record Hand : IComparable<Hand> {
-    
     public static readonly IComparer<Hand> AlternateComparer = Comparer<Hand>.Create((a, b) => a.CompareToAlternate(b));
+
+    public Hand(IList<Card> cards) {
+        if (cards.Count != 5) throw new ArgumentException("Must have 5 cards", nameof(cards));
+
+        Cards = cards.AsReadOnly();
+        Kind = CalculateKind();
+        AlternateKind = CalculateAlternateKind();
+    }
+
     public IReadOnlyList<Card> Cards { get; }
 
     public HandKind Kind { get; }
     public HandKind AlternateKind { get; }
 
+    public int CompareTo(Hand? other) {
+        if (ReferenceEquals(this, other)) return 0;
+
+        if (ReferenceEquals(null, other)) return 1;
+
+        var cmp = Kind.CompareTo(other.Kind);
+        if (cmp != 0) return cmp;
+
+        foreach (var (card, otherCard) in Cards.Zip(other.Cards)) {
+            cmp = card.CompareTo(otherCard);
+            if (cmp != 0) return cmp;
+        }
+
+        return 0;
+    }
+
     private HandKind CalculateKind() {
         var cardsWithCount = Cards.GroupBy(c => c).Select(grp => (grp.Key, grp.Count()))
             .ToDictionary(t => t.Key, t => t.Item2);
         var counts = cardsWithCount.Values;
-        if (counts.Contains(5)) {
-            return HandKind.FiveOfAKind;
-        }
+        if (counts.Contains(5)) return HandKind.FiveOfAKind;
 
-        if (counts.Contains(4)) {
-            return HandKind.FourOfAKind;
-        }
+        if (counts.Contains(4)) return HandKind.FourOfAKind;
 
         if (counts.Contains(3)) {
-            if (counts.Contains(2)) {
-                return HandKind.FullHouse;
-            }
+            if (counts.Contains(2)) return HandKind.FullHouse;
 
             return HandKind.ThreeOfAKind;
         }
 
         if (counts.Contains(2)) {
-            if (counts.Count(c => c == 2) == 2) {
-                return HandKind.TwoPair;
-            }
+            if (counts.Count(c => c == 2) == 2) return HandKind.TwoPair;
 
             return HandKind.Pair;
         }
@@ -148,13 +163,9 @@ public record Hand : IComparable<Hand> {
         var cardsWithCount = Cards.GroupBy(c => c).Select(grp => (grp.Key, grp.Count()))
             .ToDictionary(t => t.Key, t => t.Item2);
         var jokerCount = cardsWithCount.GetValueOrDefault(Card.Jack, 0);
-        if (jokerCount is 0 or 5) {
-            return Kind;
-        }
+        if (jokerCount is 0 or 5) return Kind;
 
-        if (jokerCount == 1 && Kind == HandKind.TwoPair) {
-            return HandKind.FullHouse;
-        }
+        if (jokerCount == 1 && Kind == HandKind.TwoPair) return HandKind.FullHouse;
 
         cardsWithCount.Remove(Card.Jack);
         var counts = cardsWithCount.Values;
@@ -164,18 +175,8 @@ public record Hand : IComparable<Hand> {
             4 => HandKind.FourOfAKind,
             3 => HandKind.ThreeOfAKind,
             2 => HandKind.Pair,
-            _ => throw new UnreachableException("Impossible to not have a pair with joker in hand."),
+            _ => throw new UnreachableException("Impossible to not have a pair with joker in hand.")
         };
-    }
-
-    public Hand(IList<Card> cards) {
-        if (cards.Count != 5) {
-            throw new ArgumentException("Must have 5 cards", nameof(cards));
-        }
-
-        Cards = cards.AsReadOnly();
-        Kind = CalculateKind();
-        AlternateKind = CalculateAlternateKind();
     }
 
     public static Hand Parse(string input) {
@@ -184,49 +185,18 @@ public record Hand : IComparable<Hand> {
         return new Hand(cards);
     }
 
-    public int CompareTo(Hand? other) {
-        if (ReferenceEquals(this, other)) {
-            return 0;
-        }
-
-        if (ReferenceEquals(null, other)) {
-            return 1;
-        }
-
-        var cmp = Kind.CompareTo(other.Kind);
-        if (cmp != 0) {
-            return cmp;
-        }
-
-        foreach (var (card, otherCard) in Cards.Zip(other.Cards)) {
-            cmp = card.CompareTo(otherCard);
-            if (cmp != 0) {
-                return cmp;
-            }
-        }
-
-        return 0;
-    }
-
     public int CompareToAlternate(Hand? other) {
-        if (ReferenceEquals(this, other)) {
-            return 0;
-        }
+        if (ReferenceEquals(this, other)) return 0;
 
-        if (ReferenceEquals(null, other)) {
-            return 1;
-        }
+        if (ReferenceEquals(null, other)) return 1;
 
         var cmp = AlternateKind.CompareTo(other.AlternateKind);
-        if (cmp != 0) {
-            return cmp;
-        }
+        if (cmp != 0) return cmp;
 
-        foreach (var (card, otherCard) in Cards.Select(c => c.ToAltCard()).Zip(other.Cards.Select(c => c.ToAltCard()))) {
+        foreach (var (card, otherCard) in
+                 Cards.Select(c => c.ToAltCard()).Zip(other.Cards.Select(c => c.ToAltCard()))) {
             cmp = card.CompareTo(otherCard);
-            if (cmp != 0) {
-                return cmp;
-            }
+            if (cmp != 0) return cmp;
         }
 
         return 0;
@@ -246,6 +216,8 @@ public class Gambit(Hand hand, long bet) {
 }
 
 public class Day07 : Adventer {
+    private ImmutableArray<Gambit> gambits;
+
     public Day07() {
         Bag["test"] = """
                       32T3K 765
@@ -255,8 +227,6 @@ public class Day07 : Adventer {
                       QQQJA 483
                       """;
     }
-
-    private ImmutableArray<Gambit> gambits;
 
     protected override void InternalOnLoad() {
         gambits = Input.Lines.Select(Gambit.Parse).ToImmutableArray();
