@@ -7,17 +7,12 @@ namespace AoC2023._2023;
 
 // https://en.wikipedia.org/wiki/Lagrange_polynomial
 internal class LagrangeInterpolator {
-    private List<Rational> X { get; }
-    private List<Rational> Y { get; }
-    private List<Rational> W { get; }
-
     public LagrangeInterpolator(IEnumerable<(Rational, Rational)> nodes) {
         var x = new List<Rational>();
         var y = new List<Rational>();
         foreach (var (xi, yi) in nodes) {
             x.Add(xi);
             y.Add(yi);
-            
         }
 
         X = x;
@@ -28,14 +23,16 @@ internal class LagrangeInterpolator {
     public LagrangeInterpolator(IEnumerable<(long, long)> nodes) :
         this(nodes.Select(x => (new Rational(x.Item1), new Rational(x.Item2)))) { }
 
+    private List<Rational> X { get; }
+    private List<Rational> Y { get; }
+    private List<Rational> W { get; }
+
     // thank you wikipedia
     private Rational CalculateBarycentricWeight(int j) {
         var x = X[j];
         var w = Rational.One;
         for (var m = 0; m < X.Count; m++) {
-            if (j == m) {
-                continue;
-            }
+            if (j == m) continue;
 
             w /= x - X[m];
         }
@@ -49,9 +46,7 @@ internal class LagrangeInterpolator {
 
     public Rational Interpolate(Rational x) {
         var idx = X.FindIndex(xi => xi == x);
-        if (idx != -1) {
-            return Y[idx];
-        }
+        if (idx != -1) return Y[idx];
 
         var numerator = Rational.Zero;
         var denominator = Rational.Zero;
@@ -70,17 +65,12 @@ internal class LagrangeInterpolator {
 }
 
 internal class LagrangeInterpolator<T> where T : INumber<T> {
-    private List<T> X { get; }
-    private List<T> Y { get; }
-    private List<T> W { get; }
-
     public LagrangeInterpolator(IEnumerable<(T, T)> nodes) {
         var x = new List<T>();
         var y = new List<T>();
         foreach (var (xi, yi) in nodes) {
             x.Add(xi);
             y.Add(yi);
-            
         }
 
         X = x;
@@ -91,14 +81,16 @@ internal class LagrangeInterpolator<T> where T : INumber<T> {
     public LagrangeInterpolator(IEnumerable<(long, long)> nodes) :
         this(nodes.Select(x => (T.CreateChecked(x.Item1), T.CreateChecked(x.Item2)))) { }
 
+    private List<T> X { get; }
+    private List<T> Y { get; }
+    private List<T> W { get; }
+
     // thank you wikipedia
     private T CalculateBarycentricWeight(int j) {
         var x = X[j];
         var w = T.One;
         for (var m = 0; m < X.Count; m++) {
-            if (j == m) {
-                continue;
-            }
+            if (j == m) continue;
 
             w /= x - X[m];
         }
@@ -112,9 +104,7 @@ internal class LagrangeInterpolator<T> where T : INumber<T> {
 
     public T Interpolate(T x) {
         var idx = X.FindIndex(xi => xi == x);
-        if (idx != -1) {
-            return Y[idx];
-        }
+        if (idx != -1) return Y[idx];
 
         var numerator = T.Zero;
         var denominator = T.Zero;
@@ -134,12 +124,13 @@ internal class LagrangeInterpolator<T> where T : INumber<T> {
 
 internal class Day09Problem(IReadOnlyList<long> numbers) {
     public IReadOnlyList<long> Numbers { get; } = numbers;
+
     public Func<long, long> GetSimpleInterpolator() {
         var c = Numbers[0];
         var degree = 0;
         var currentLevel = Numbers;
         while (!currentLevel.AllEqual()) {
-            currentLevel = currentLevel.Zip(currentLevel.Skip(1)).Select((a) => a.Second - a.First).ToArray();
+            currentLevel = currentLevel.Zip(currentLevel.Skip(1)).Select(a => a.Second - a.First).ToArray();
             degree++;
         }
 
@@ -147,19 +138,21 @@ internal class Day09Problem(IReadOnlyList<long> numbers) {
 
         return x => coefficient * x.Pow(degree) + c;
     }
-    
+
     public Func<long, long> GetLagrangeInterpolator() {
         var interp = new LagrangeInterpolator(Numbers.Select((x, i) => ((long)i, x)));
         return x => (long)interp.Interpolate(x).WholePart;
     }
 
     public Func<long, long> GetLagrangeInterpolator<T>() where T : INumber<T> {
-        var interp = new LagrangeInterpolator<T>(Numbers.Select((x, i) => ((long) i, x)));
+        var interp = new LagrangeInterpolator<T>(Numbers.Select((x, i) => ((long)i, x)));
         return x => long.CreateChecked(Math.Round(decimal.CreateChecked(interp.Interpolate(x))));
     }
 }
 
 public class Day09 : Adventer {
+    private IReadOnlyList<Day09Problem> problems;
+
     public Day09() {
         Bag["test"] = """
                       0 3 6 9 12 15
@@ -167,8 +160,6 @@ public class Day09 : Adventer {
                       10 13 16 21 30 45
                       """;
     }
-    
-    private IReadOnlyList<Day09Problem> problems;
 
     [MemberNotNull(nameof(problems))]
     protected override void InternalOnLoad() {
@@ -179,23 +170,17 @@ public class Day09 : Adventer {
 
     protected override object InternalPart1() {
         long total = 0;
-        foreach (var p in problems) {
-            total += p.GetLagrangeInterpolator()(p.Numbers.Count);
-            // total += p.GetSimpleInterpolator()(p.Numbers.Count);
-            // total += p.GetLagrangeInterpolator<decimal>()(p.Numbers.Count);
-        }
-
+        foreach (var p in problems) total += p.GetLagrangeInterpolator()(p.Numbers.Count);
+        // total += p.GetSimpleInterpolator()(p.Numbers.Count);
+        // total += p.GetLagrangeInterpolator<decimal>()(p.Numbers.Count);
         return total;
     }
 
     protected override object InternalPart2() {
         long total = 0;
-        foreach (var p in problems) {
-            total += p.GetLagrangeInterpolator()(-1);
-            // total += p.GetSimpleInterpolator()(-1);
-            // total += p.GetLagrangeInterpolator<decimal>()(-1);
-        }
-
+        foreach (var p in problems) total += p.GetLagrangeInterpolator()(-1);
+        // total += p.GetSimpleInterpolator()(-1);
+        // total += p.GetLagrangeInterpolator<decimal>()(-1);
         return total;
     }
 }
