@@ -21,31 +21,67 @@
 using System.Runtime.CompilerServices;
 using AoC.Support;
 using QuikGraph;
-using QuikGraph.Graphviz;
 
 namespace AoC2023;
 
 public class Day20 : Adventer {
     public enum Pulse {
         Low = 0,
-        High = 1,
+        High = 1
+    }
+
+    private Problem problem;
+
+    public Day20() {
+        Bag["test"] = """
+                      broadcaster -> a, b, c
+                      %a -> b
+                      %b -> c
+                      %c -> inv
+                      &inv -> a
+                      """;
+    }
+
+    protected override void InternalOnLoad() {
+        problem = new Problem(Input.Lines);
+    }
+
+    protected override object InternalPart1() {
+        var machine = problem.BuildMachine();
+        for (var i = 0; i < 1000; i++) machine.PushButton();
+        return machine.PulseCounters.Product(c => (long)c.Item2);
+    }
+
+    protected override object InternalPart2() {
+        var machine = problem.BuildMachine();
+        var firstObservedHigh = new Dictionary<string, long>();
+        var pushes = new[] { 0L };
+        machine.ObserveMessage += msg => {
+            if (msg is { Datum: Pulse.High, Recipient.Id: "jz" }) firstObservedHigh.TryAdd(msg.Sender.Id, pushes[0]);
+        };
+        while (firstObservedHigh.Count < 4) {
+            pushes[0]++;
+            machine.PushButton();
+        }
+
+        return firstObservedHigh.Values.Aggregate(MathUtils.Lcm);
     }
 
     public abstract class Module(string id) : IEquatable<Module> {
-        public bool Equals(Module? other) {
-            if (ReferenceEquals(null, other)) {
-                return false;
-            }
+        public string Id { get; } = string.Intern(id);
 
-            if (ReferenceEquals(this, other)) {
-                return true;
-            }
+        public abstract string Prefix { get; }
+
+        public bool Equals(Module? other) {
+            if (ReferenceEquals(null, other)) return false;
+
+            if (ReferenceEquals(this, other)) return true;
 
             return ReferenceEquals(Id, other.Id);
         }
 
         public override bool Equals(object? obj) {
-            return ReferenceEquals(this, obj) || obj is Module other && Equals(other);
+            return ReferenceEquals(this, obj) || (obj is Module other && Equals(other));
         }
 
         public override int GetHashCode() {
@@ -60,28 +96,28 @@ public class Day20 : Adventer {
             return !Equals(left, right);
         }
 
-        public string Id { get; } = string.Intern(id);
         public abstract Pulse? HandleInput(Module sender, Pulse input);
-
-        public abstract string Prefix { get; }
 
         public override string ToString() {
             return $"{Prefix}{Id}";
         }
-    };
+    }
 
     public class Broadcast() : Module("broadcaster") {
-        public override Pulse? HandleInput(Module _, Pulse input) => input;
         public override string Prefix => "";
+
+        public override Pulse? HandleInput(Module _, Pulse input) {
+            return input;
+        }
     }
 
     public class FlipFlop(string id) : Module(id) {
         public bool Powered { get; set; }
 
+        public override string Prefix => "%";
+
         public override Pulse? HandleInput(Module _, Pulse input) {
-            if (input == Pulse.High) {
-                return null;
-            }
+            if (input == Pulse.High) return null;
 
             if (Powered) {
                 Powered = false;
@@ -91,13 +127,13 @@ public class Day20 : Adventer {
             Powered = true;
             return Pulse.High;
         }
-
-        public override string Prefix => "%";
     }
 
     public class Conjunction(string id, IEnumerable<string> sources) : Module(id) {
+        private Dictionary<string, Pulse> Inputs { get; } = sources.ToDictionary(s => s, _ => Pulse.Low,
+            EqualityComparer<string>.Create(ReferenceEquals, RuntimeHelpers.GetHashCode));
 
-        private Dictionary<string, Pulse> Inputs { get; } = sources.ToDictionary(s => s, _ => Pulse.Low, EqualityComparer<string>.Create(ReferenceEquals, RuntimeHelpers.GetHashCode));
+        public override string Prefix => "&";
 
         private void ObserveInput(Module sender, Pulse input) {
             Inputs[sender.Id] = input;
@@ -106,45 +142,64 @@ public class Day20 : Adventer {
         public override Pulse? HandleInput(Module sender, Pulse input) {
             ObserveInput(sender, input);
 
-            foreach (var pulse in Inputs.Values) {
-                if (pulse == Pulse.Low) {
+            foreach (var pulse in Inputs.Values)
+                if (pulse == Pulse.Low)
                     return Pulse.High;
-                }
-            }
 
             return Pulse.Low;
         }
-
-        public override string Prefix => "&";
     }
-    
+
     public class Dummy(string id) : Module(id) {
+        public override string Prefix => "";
+
         public override Pulse? HandleInput(Module sender, Pulse input) {
             return null;
         }
-
-        public override string Prefix => "";
     }
 
     public readonly record struct Message(Module Sender, Pulse Datum, Module Recipient) {
         public override string ToString() {
             return $"{Sender} -{Datum}-> {Recipient}";
         }
-
     }
 
     public class Machine {
-
-        private readonly ulong[] pulseCounters = { 0ul, 0ul };
-        private readonly Dictionary<string, Module> modules = new();
-        private readonly Dictionary<Module, List<Module>> moduleGraph = new(); 
         private readonly Module broadcaster;
         private readonly Queue<Message> messageQueue = new();
+        private readonly Dictionary<Module, List<Module>> moduleGraph = new();
+        private readonly Dictionary<string, Module> modules = new();
+
+        private readonly ulong[] pulseCounters = { 0ul, 0ul };
+
+        public Machine(IBidirectionalGraph<string, Edge<string>> blueprint, IDictionary<string, char> idToPrefix) {
+            broadcaster = new Broadcast();
+            foreach (var vert in blueprint.Vertices) {
+                var vertex = string.Intern(vert);
+                if (!idToPrefix.TryGetValue(vertex, out var prefix)) prefix = default;
+                var module = prefix switch {
+                    '%' => new FlipFlop(vertex),
+                    '&' => new Conjunction(vertex, blueprint.InEdges(vertex).Select(e => e.Source)),
+                    _ when vertex == "broadcaster" => broadcaster,
+                    _ => new Dummy(vertex)
+                };
+
+                modules[vertex] = module;
+            }
+
+            foreach (var vertex in blueprint.Vertices)
+                moduleGraph[modules[vertex]] = blueprint.OutEdges(vertex).Select(e => modules[e.Target]).ToList();
+            // flipFlops.Sort(((flop, flipFlop) => string.Compare(flop.Id, flipFlop.Id, StringComparison.Ordinal)));
+        }
+
+        private bool Quiescent => messageQueue.Count == 0;
+
+        public IEnumerable<(Pulse, int)> PulseCounters => pulseCounters.Select((c, i) => ((Pulse)i, (int)c));
 
         public event Action<Message>? ObserveMessage;
 
         private void ObservePulse(Pulse pulse) {
-            pulseCounters[(int) pulse]++;
+            pulseCounters[(int)pulse]++;
         }
 
         public void PushButton() {
@@ -152,47 +207,17 @@ public class Day20 : Adventer {
             Quiesce();
         }
 
-        private bool Quiescent => messageQueue.Count == 0;
-
         private void Quiesce() {
             while (!Quiescent) {
                 var msg = messageQueue.Dequeue();
                 ObserveMessage?.Invoke(msg);
                 ObservePulse(msg.Datum);
                 var output = msg.Recipient.HandleInput(msg.Sender, msg.Datum);
-                if (output == null) {
-                    continue;
-                }
+                if (output == null) continue;
 
-                foreach (var destination in moduleGraph[msg.Recipient]) {
+                foreach (var destination in moduleGraph[msg.Recipient])
                     messageQueue.Enqueue(new Message(msg.Recipient, output.Value, destination));
-                }
             }
-        }
-
-        public IEnumerable<(Pulse, int)> PulseCounters => pulseCounters.Select((c, i) => ((Pulse) i, (int) c));
-
-        public Machine(IBidirectionalGraph<string, Edge<string>> blueprint, IDictionary<string, char> idToPrefix) {
-            broadcaster = new Broadcast();
-            foreach (var vert in blueprint.Vertices) {
-                var vertex = string.Intern(vert);
-                if (!idToPrefix.TryGetValue(vertex, out var prefix)) {
-                    prefix = default(char);
-                }
-                var module = prefix switch {
-                    '%' => new FlipFlop(vertex),
-                    '&' => new Conjunction(vertex, blueprint.InEdges(vertex).Select(e => e.Source)),
-                    _ when vertex == "broadcaster" => broadcaster,
-                    _ => new Dummy(vertex),
-                };
-
-                modules[vertex] = module;
-            }
-
-            foreach (var vertex in blueprint.Vertices) {
-                moduleGraph[modules[vertex]] = blueprint.OutEdges(vertex).Select(e => modules[e.Target]).ToList();
-            }
-            // flipFlops.Sort(((flop, flipFlop) => string.Compare(flop.Id, flipFlop.Id, StringComparison.Ordinal)));
         }
     }
 
@@ -208,24 +233,22 @@ public class Day20 : Adventer {
                 var id = prefix switch {
                     '%' => span[1..arrow].Trim().ToString(),
                     '&' => span[1..arrow].Trim().ToString(),
-                    _   => span[..arrow].Trim().ToString(),
+                    _ => span[..arrow].Trim().ToString()
                 };
                 id = string.Intern(id);
                 var destStart = arrow + 2;
                 var current = span[destStart..].Trim();
                 var destEnd = current.IndexOf(',');
-                if (destEnd == -1) {
-                    destEnd = current.Length;
-                }
+                if (destEnd == -1) destEnd = current.Length;
                 while (current.Length > 0) {
                     var dest = current[..destEnd].Trim().ToString();
                     blueprint.AddVerticesAndEdge(new Edge<string>(id, dest));
                     destEnd = current.IndexOf(',');
-                    if (destEnd == -1) {
-                        destEnd = current.Length-1;
-                    }
+                    if (destEnd == -1) destEnd = current.Length - 1;
                     current = current[(destEnd + 1)..].Trim();
-                };
+                }
+
+                ;
 
                 idToPrefix[id] = prefix;
             }
@@ -234,46 +257,5 @@ public class Day20 : Adventer {
         public Machine BuildMachine() {
             return new Machine(blueprint, idToPrefix);
         }
-
-    }
-
-    public Day20() {
-        Bag["test"] = """
-                      broadcaster -> a, b, c
-                      %a -> b
-                      %b -> c
-                      %c -> inv
-                      &inv -> a
-                      """;
-    }
-
-    private Problem problem;
-    protected override void InternalOnLoad() {
-        problem = new Problem(Input.Lines);
-    }
-
-    protected override object InternalPart1() {
-        var machine = problem.BuildMachine();
-        for (var i = 0; i < 1000; i++) {
-            machine.PushButton();
-        }
-        return machine.PulseCounters.Product(c => (long)c.Item2);
-    }
-
-    protected override object InternalPart2() {
-        var machine = problem.BuildMachine();
-        var firstObservedHigh = new Dictionary<string, long>();
-        var pushes = new[] { 0L };
-        machine.ObserveMessage += msg => {
-            if (msg is { Datum: Pulse.High, Recipient.Id: "jz" }) {
-                firstObservedHigh.TryAdd(msg.Sender.Id, pushes[0]);
-            }
-        };
-        while (firstObservedHigh.Count < 4) {
-            pushes[0]++;
-            machine.PushButton();
-        }
-
-        return firstObservedHigh.Values.Aggregate(MathUtils.Lcm);
     }
 }
