@@ -1,4 +1,5 @@
 ﻿#region license
+
 // AoC2023 - AoC2023 - Day22.cs
 // Copyright (C) 2023 Nicholas
 // 
@@ -14,16 +15,12 @@
 // 
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
-#endregion
 
-using NetTopologySuite.Mathematics;
+#endregion
 
 namespace AoC2023._2023;
 
-using Vertex = Vector3D;
-
-public class Day22: Adventer {
-    
+public class Day22 : Adventer {
     protected override object InternalPart1() {
         throw new NotImplementedException();
     }
